Refactor import paths and add status field to Project model
