+++ apps/main_page/app/components/ShowcaseItem.tsx
@@ -1,13 +1,13 @@
 "use client"
 import Image from "next/image";
 import { motion } from "framer-motion"
-import { Project } from '../model/project';
+import { Project } from '../../model/project';
 import { client } from "../../sanity/lib/client";
 import imageUrlBuilder from '@sanity/image-url'
 
 export default function ShowcaseItem({ project }: { project: Project }): JSX.Element {
   const builder = imageUrlBuilder(client)
-  return <motion.a className="col-12 col-md-6 col-lg-4 text-decoration-none" href={project.url} target="_blank">
+  return <motion.a className={`col-12 col-md-6 col-lg-4 text-decoration-none ${project.status == "live" ? "live" : "upcoming"}`} href={project.url} target="_blank">
     <div className="card showcase-card">
       <motion.div className="image-wrapper card-image-wrapper" whileHover={{ scale: 1.02, translateY: -10 }} transition={{
         ease: "easeInOut",
